pyproject.toml
[tool.poetry]
name = "app-revolvedor"
version = "0.1.0"
description = ""
authors = ["Daniel Donateli <danieldonateli@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
